`ConnectionLimitReachedException` should not trigger RRLB health-checking (#2457)
Motivation:

`ConnectionLimitReachedException` should implement a `RetryableException` contract
and should not mark the remote host unhealthy bcz it's a local limit that does not
affect the host's actual healthiness. Hosts are not marked as "unhealthy" if there
are open connections to them, but there is currently a risk to changes host's
state to "unhealthy" from the cold start, if we wait for allowed connections to
open.

Modifications:

- Enhance `markUnhealthy` logic inside `RoundRobinLoadBalancer` to ignore
`ConnectionLimitReachedException` for counting towards the threshold;
- Test that `ConnectionLimitReachedException` doesn't move a host to
"unhealthy" state;
- Add `RetryableException` marker for `ConnectionLimitReachedException`;
- Clarify `LimitingConnectionFactoryFilter`'s javadoc about recommended
exception;
- Implement `newConnectionRefusedException` inside `MaxConnectionsLimiter`
to include the current limit at the end of exception message;

Results:

1. `ConnectionLimitReachedException` doesn't change the state of the host.
2. `ConnectionLimitReachedException` is marked as `RetryableException`.
3. `LimitingConnectionFactoryFilter.withMax(int)` includes its limit in
the exception message.