Provide common HTTP Service interface for all API variants. (#2456)
Motivation:

Right now four different Service API variants are provided, all
with their async and sync listen* overloads. If the service
type is known at compile time those overloads suffice, but in
situations where the type is only determined at runtime (for
example when DI is used) the user needs to write verbose code
that can be provided by ServiceTalk and improve the developer
experience.

Modifications:

This changeset introduces a new "Service" interface in the
http package and it now acts as a parent for the four different
service implementations. The HttpServerBuilder gets two more
overloads (blocking and non-blocking) to accept the Service
type and then internally performs the runtime type checking
and delegates to the appropriate methods.

Result:

Better developer experience when the service type is not known
at compile time, for example when dependency injection is used.