Make RouteExecutionStrategyFactory optional on Builder
Motivation:

Now that route building has been made lazy in a previous commit,
the RouteExecutionStrategyFactory can be moved to an optional
builder argument and therefore clean up the API.

Modifications:

In addition to making RouteExecutionStrategyFactory available
on the Builder, the corresponding constructors have been marked
as Deprecated since they are not needed anymore and can be
removed in a future version.

NOTE: the test only asserts that the custom factory is called,
due to a different potential bug identified we cannot assert
that custom offloading parameters are actually applied (yet).
Once resolved, this will be implemented in a follow-up
commit.

Result:

Simpler user-facing API when building GRPC services.