Deprecate ConnectionAcceptor(+ Factory) "append"
Motivation:

At the moment the user experience on how to apply multiple
ConnectionAcceptorFactories is confusing because there are
many different ways on how to perform this task.

In order to improve the ambiguity there should only be one
common way to append multiple acceptor factories.

Modifications:

This changeset deprecates the append methods on both the
ConnectionAcceptor and the ConnectionAcceptorFactory. A
new ConnectionAcceptorFactoryAppender is added which is used
internally and in tests, but can be used by a user who needs
to port functionality forward when relying on the old
behavior.

All internal use of the deprecated APIs has been replaced
with the helper class in preparation for later removal.

Result:

The primary API to chain the acceptors should be through
the regular HttpServerBuilder which is now expressed through
deprecation and later removal of the other APIs.