Introduce GrpcExecutionStrategies#offloadNone (#2478)
* Introduce GrpcExecutionStrategies#offloadNone

Motivation:

To ensure partiy with HttpExecutionStrategies it makes sense to
introduce "offloadNone" as a static utility method instead of
only making it accessible through a builder.

Modifications:

This changeset adds the static "offloadNone" utility method in
GrpcExecutionStrategies and clarifies the javadoc of "offloadNever"
(which is deprecated) on alternative usage. It also modifies
the Http conversions to include the new strategy explicitly.

Result:

Feature parity between GRPC and HTTP on the "offloadNone" execution
strategy creation.