Introduce Early and Late ConnectionAcceptors
Motivation:

The currently available ConnectionAcceptor has the
following shortcomings which are attempted to be
addressed with this changeset:

 - Not being able to reject connections before the
   TLS handshake is performed.
 - Complex semantics how to append / chain multiple
   Acceptors together for the same service.

Modifications:

The early and late stages in accepting a connection
are now broken out into two separate interfaces which
accept different amount of information - depending on
what is available in the current stage of the connection.

Each acceptor now only knows itself and the appending
of multiple acceptors is performed by the HttpServerBuilder,
simplifying appending multiple acceptors from a user
perspective.

Since they are different interfaces, different offloading
strategies can be put in place for each - still defaulting
to full offloading.

Note: this changeset does not deprecate the "old"
acceptor, this will be done in a follow-up commit.

Result:

Improved functionality and user experience when needing to
reject connections early as well as chaining multiple acceptors
together.