Revert "Preserve the caller's stacktrace for blocking API (#2420)" (#2488)
This reverts commit 8b41bf4090fdc0503b11df855d98e5b7b0a4c5bd.

Motivation: 

There are cases when Netty's DNS resolver may put an exception into the cache and return it multiple times. In this case, we keep adding suppressed exceptions to the same instance and it leads to a memory leak and increased logging output.
Let's revert this feature before the upcoming release and revisit later when we figure out the approach for Netty/ST DNS resolver.