Deprecate ConnectionAcceptor (Factory)
Motivation:

A previous commit added new functionality (early and late acceptor)
which supersede the current ConnectionAcceptor functonality.

Modifications:

Deprecate the ConnectionAcceptor as well as the outer
ConnectionAcceptorFactory and the high level usage API in
HttpServerBuilder. The corresponding replacement is the new
LateConnectionAcceptor which differs slightly in its implementation
but is simpler to use than this one when chaining multiple ones.

Result:

Deprecating the current ConnectionAcceptor in favor for the new
early and late acceptors.

Add default implementation