Fix some broken {source-root} references in documentation (#2493)
Motivation:
Some of the documentations .adoc files are missing a definition for
source-root but use them in hyperlinks. This results in broken links.
This happens easily because each file that references source-root
needs the definition.

Modifications:
Add a script to check all the .adoc files for missing source-root
definitions. We also run the script with the -r flag to fix up the broken
files. To avoid regressions, the script is run as part of the Github
PR Quality workflow.

Result:
Files are fixed and it's easier to check this in the future.