Implement TLS Certificate Compression Support (#2497)
Motivation:

RFC 8879 defines a TLS extension called "TLS Certificate
Compression" which allows to compress certificates and
possibly reduce the amount of data exchanged during a
TLS handshake.

Since netty already added support for it when used together
with BoringSSL (netty-tcnative), ServiceTalk can benefit
from a higher-level integration.

Modifications:

This changeset implements ServiceTalk API for Certificate
Compression and hooks it up with the Netty APIs internally.

The main entry point for the user is through the SslConfigBuilder
where they can set the supported algorithms (for now only ZLIB)
for both client and server use cases. We also support both the
compression and decompression depending on the use case.

At the moment it is not possible to define custom algorithms,
although the API has been design with this use-case in mind to
open it up later if needed.

ZlibOpenSslCertificateCompressionAlgorithm performs the actual
compression and decompression and is called from the netty layer
when doing so.

NOTE: since certificate compression is a "best effort" optimization,
there is no failure if the feature is not supported or not available
at runtime. If needed, inspection through a TransportObserver can
determine if the number of bytes exchanged during a handshake are
decreased or not.

NOTE: at the time of merge, the BoringSSL implementation that is
used in netty does not support client-side compression. This does
not affect the high level API usage in ServiceTalk, but can be 
surprising when testing this feature.

Result:

TLS certificate compression with ZLIB can now be enabled and lead
to less network overhead when performing the TLS handshake.