Use SslProvider.isOptionSupported to determine cert compression support
Motivation:

Netty recently added a getter to SslProvider which allows us to check
if certain options are supported or not.

Modifications:

This changeset changes the SslContextFactory to use this new getter
and explicitly check for
OpenSslContextOption.CERTIFICATE_COMPRESSION_ALGORITHMS before enabling
certificate compression.

Result:

More reliable check if certificate compression is supported or not.