Implement explicit, customizable DNS maxTTL
Motivation:

At the moment it is possible to configure a minTTL for DNS
resolution, but maxTTL is not configured (the cache uses
Integer.MAX_INT) and the actual maxTTL for resolution is
applied to whatever the DNS entry returns from the server.

To provide more customizations as well as being able to
explicitly "cap" a TTL at a maximum value, it makes sense
to allow setting an explicit maximum TTL.

Modifications:

This changeset introduces a configurable maxTTL option on the
DefaultDnsServiceDiscovererBuilder with a sensible default
setting. This value then:

 - Propagates to netty's DefaultDnsCache as the maximum value

as well as

 - Caps the returned TTL from the server to this value and
   logs a INFO level message if it has been capped (since reaching
   the MAX is really not expected in a well configured environment
   and can point to issues the actual deployed system).

Result:

Configurable DNS maxTTL as well as TTL capping for the DNS records
to provide better safety at runtime.