Implement DnsServiceDiscovererBuilderProvider (#2511)
Motivation:

Similar to the related functionality in GRPC and HTTP servers,
it can be useful to configure the DNS service discoverer also
through a ServiceProvider.

Modifications:

This changeset introduces ServiceProvider functionality on top
of the DefaultDnsServiceDiscovererBuilder. To do this, a new
interface is introduced (to also provide a Delegate implementation)
as well as a DnsServiceDiscoverers factory class to construct a
builder with an ID.

Note that uniqueness on this ID is not enforced by ServiceTalk,
but is something to recommend to the application developer when
using this mechanism. The default construct is marked as
deprecated and will use a UUID as the ID.

Result:

It is now possible to configure the DNS service discoverer through
the ServiceProvider mechanism.