Disable cache by default for DNS ServiceDiscoverer
Motivation:

When we poll DNS in the background the cache is not important because
we always schedule the next query after cache expires. It only helps
with concurrent resolutions of the same address, which in practice
should be minimal if users properly reuse clients. Disadvantage of
having a cache is that if it gets poisoned with invalid or stale entries
there is no way to clear the cache. Cancelling the events publisher and
re-subscribing doesn't help because re-subscribe always hits the cache.
See #2514.

Modifications:

- Disable cache by default;
- Provide API to opt-in for caching (can be useful if users resolve per
new connection instead of polling in the background);
- Deprecate `minTTL`, remove `maxTTL` builder methods, introduce
`ttl(min, max, cache)` instead;
- Invoke `ttlCache.prepareForResolution(name)` only for scheduled queries,
keep it unchanged when cancel/re-subscribe to correctly offset ttl;
- Ignore empty list inside `MinTtlCache.get(...)`;
- Make `DefaultDnsClient` logging more consistent;
- Enhance testing;

Result:

Caching is disabled by default. Polling is driven by TTL. In case of
re-subscribe, we always send a new query. Users have API to configure
min/max polling intervals and caching.