Implement RoundRobinLoadBalancerBuilderProvider
Motivation:

It can be useful to configure the RoundRobinLoadBalancer
through a ServiceProvider, similar to other components
already present in ServiceTalk.

Modifications:

This changeset introduces ServiceProvider functionality
on top of the RoundRobinLoadBalancerBuilder. To do this,
the builder has been refactored into an interface and
a delegate implementation is also introduced.

Result:

It is now possible to configure the RoundRobinLoadBalancer
through the ServiceProvider mechanism.