Extend DnsServiceDiscovererObserver functionality (#2521)
Motivation:

To enhance the utility of the DnsServiceDiscovererObserver a couple
additional APIs can be provided:

 - The newly introduced ServiceDiscoverer ID is brought in scope
 - Further lifecycle callbacks are added to the DnsDiscoveryObserver

Modifications:

This changeset adds the two enhancements mentioned in the motivation
in a backwards-compatible way. In the next minor released the interface
can be cleaned up.

Tests have been enhanced to test the newly added functionality.

Result:

More functionality around the DnsServiceDiscovererObserver.