Enhance `SniTest` to verify that SNI is present in `SSLSession`
Motivation:

It may be necessary for users to read SNI used for the SSL handshake to
apply different policies. Let's make sure the SNI is always present
in `SSLSession`.

Modifications:

- Enhance `SniTest` to validate that `getRequestedServerNames()`
contains value regardless of the used `SslProvider` and that
`SNIHostName` matches the expected value;

Result:

Better test coverage for SNI use-case.