Allow to wrap Executor with custom TimeSource
Motivation:

When an Executor needs to wrapped with a custom TimeSource
(i.e for testing) it is possible, but not very obvious or
user friendly at the moment.

Modifications:

This changeset introduces a factory method to help delegate
calles from a provided executor and also use a custom TimeSource.

Result:

It is now easier to use an executor with a custom TimeSource if
needed.