Use an unbounded queue for fixed size Executors
Motivation:

It's very difficult to properly use an Executor that has a fixed size
thread pool because the current queue is a SynchronousQueue which will
result in rejecting work if there isn't a free thread to service it.
This also runs contrary to the patterns used in the JDK for fixed size
thread pools which default to the unbounded LinkedBlockingQueue.

Modifications:

Use the unbounded LinkedBlockingQueue for Executors with fixed size.

Result:

Less surprising behavior of the fixed size Executors.