Allow null values with FormUrlEncodedHttp[De]Serializer
Motivation:

At the moment the behavior around null (and to an extent empty)
values is a little bit unflexible. ServiceTalk makes the decision
to ignore null values on encoding and converting them to the same
value as an empty string on decoding.

Since there is no spec which says that null values are disallowed,
it is better if ServiceTalk is flexible and lets the user decide
if they want to ignore values (which they can do by checking and
then just not encoding them in the first place).

Modifications:

This changeset modifies behavior as follows.

On the encoding side:
 - null values are now allowed and encoded as "just" the key,
   instead of not being encoded at all.
   (Distinguished without an "=" compared to an empty string)

On the decoding side:
 - Null values are now not turned into an empty string, but if
   just the key is present turned into "null" - and if the "="
   is present, an empty string is decoded.

Note: this also implicitly changes the resulting HttpQuery behavior
when calling lazyParseQueryString on HttpRequestMetaData, since
it uses the same functionality underneath. Since the "get" method
is marked as @Nullable already, there is no illegal behavior observed.