Fix race in TimeoutDemandPublisher (#2578)
Motivation:

There is a race condition in TimeoutDemandPublisher which can
result in a erroneous timeout if we have a call to .request(n) racing
with a .onNext(X) call. The result is an erroneous timer that will
timeout the publisher despite demand being present.

Modifications:

Leverage the invariant of the system that if the demand is 0 there
must be a timer set: if we transition from 0 -> n and find the timer
to be null, we must be experiencing the race. The solution is to
wait for the thread in .onNext(X) to finish setting the timer.

Result:

One less bug.