Jersey router: async endpoints don't propagate AsyncContext
Motivation:

When users use `AsyncContext` inside endpoints that operate with async
primitives, the context is not propagated as expected on the response
path.

Modifications:

- `EndpointEnhancingRequestFilter`: add `shareContextOnSubscribe()` when
we subscribe to `Single<Response>` and when we use `flatMap` to
`handleContainerResponse`;
- Add `AsyncContextTest` and `AsyncContextResources` to verify expected
behavior for different endpoints;
- Make more `AsyncContextHttpFilterVerifier` utilities public to share
code for `AsyncContext` verification;
- Avoid using deprecated `Matchers.isEmptyString()`;
- Remove outdated exclusion from `testFixtures-exclusions.xml`;

Result:

Jersey endpoints correctly propagate `AsyncContext`.