`KeepAliveManager`: force close the channel in case of any issues
Motivation:

1. If we failed to write a `PING` or `GO_AWAY` frame, the channel is
already corrupted and we should not expect a gentle FIN exchange, simply
close the channel.
2. If we shutdown output, wait for input shutdown only for a reasonable
time frame to prevent a resource leak because network can lose a FIN
back.

Modifications:
- Close the channel immediately if `close0` is invoked with an
exception;
- Do `SslHandler.closeOutbound()` only for a duplex channel;
- Schedule a timer for `pingAckTimeoutNanos` to wait for input shutdown
and force close if it's not received;

Result:

HTTP/2 connections don't hang indefinitely if the remote peer does not
reply with a FIN back or if FIN was lost.