`RetryingHttpRequesterFilter`: don't wait for LB if it's unhealthy (#2648)
Motivation:

Today, `NoAvailableHostException` can be thrown when either LB is not
ready yet or when it's already ready, but all hosts turned into
unhealthy state. In this case, there is no need to wait for "ready"
event, because the onHostsAvailable completable is already completed.
There is a risk to never exit from the retry loop if
`RetryingHttpRequesterFilter` is configured with
`maxTotalRetries(Integer.MAX_VALUE)`.

Modifications:

- Introduce `NoActiveHostException` to avoid interacting with
`LoadBalancerReadySubscriber` when LB turns into unhealthy state;
- Add tests to verify expected behavior;
- Adjust `RoundRobinLoadBalancerTest`;

Result:

`RetryingHttpRequesterFilter` doesn't involve
`LoadBalancerReadySubscriber` when LB is unhealthy.