Mark connection as closing on exception caught
Motivation:

`NettyChannelPublisher` will force closure of the `Channel` in case of
exception after the cause is propagated to users. In case users don't
have offloading, there is a risk to retry on the same IO thread.
We should notify `LoadBalancer` that this connection is closing to avoid
retrying on the same connection.

Modifications:

- Invoke `connection.notifyOnClosing()` inside `exceptionCaught` before
notifying `NettyChannelPublisher` about an error;
- Enhance `ReserveConnectionTest` to make sure reserved connection is
closed in case of `IOException`;
- Add `ConnectionClosedAfterIoExceptionTest` to make sure retries use
a different connection in case of `IOException` and the failed
connection gets closed;

Result:

Retries always select a different connection if existing connection
observes an exception.