Clean up client response payloads on error to free up connections
Motivation
----------
When a client makes a request and during response processing an
exception is thrown, the exception bubbles up to the client but
the message payload body is then not available to be consumed.

In this case, the underlying pooled connection will not be freed
up and a new connection is created.

Modifications
-------------
By tracking the message payloads when responses are bubbled up
through the filter chain, we can detect if an error happens
and if detected the message payload body is being proactively
drained.

This will free up the connection again so it can be reused by
a subsequent request.

Result
------
More correct handling of message payload bodies if an exception
happens during response filter processing.