Calculate initial concurrency based on resulting connection protocol
Motivation:

`RequestConcurrencyController` is always applied as the top level
wrapper of the connection (after all connection-level filters) before
returning it back to the `LoadBalancer`. As the result, there is no
need to make an earlier decision for `initialConcurrency` value at
`AbstractLBHttpConnectionFactory` implementation (when the final
protocol may be unknown) because it can be inferred based on the
resulting connection protocol at the time a new
`ReservableRequestConcurrencyController` is created. Early
initialization as it's now makes it harder to switch protocols after
proxy connect.

Modifications:

- Remove abstract method
`AbstractLBHttpConnectionFactory.newConcurrencyController`;
- Compute `initialConcurrency` based on connection's protocol and
`ReadOnlyHttpClientConfig`;
- Update `ReservableRequestConcurrencyControllersTest`;

Result:

No need to pre-initialize `initialConcurrency` value in
`AbstractLBHttpConnectionFactory` implementations as it will be computed
based on the connection's protocol at the time it's required.