Introduce defaultHttp(s)Port for MultiAddressHttpClientBuilder
Motivation
----------
At the moment, if no port number is explicitly provided, the
default HTTP and HTTPS ports are used (80 and 443). In some
environments it can be useful (i.e. if a client is always talking
to a high port number) to not have to specify it on every
request.

Modifications
-------------
This changeset introduces two new builder options to allow customizing
the defaultHttpPort as well as the defaultHttpsPort and applying it
if no explicit port is provided on the request itself.

Result
------
It is now possible to not specify the port number on every request
if they diverge from the default 80 or 443 port numbers.