Extract inner types from RoundRobinLoadBalancer
Motivation:

The RoundRobinLoadBalancer contains a bunch of code that is doing a
good deal more than load balancing, namely it's doing pooling, circuit
breaking, and other things. This both makes the implementation hard to
read and reason about and makes the inner pieces difficult to reuse for
future load balancer implementation.

Modifications:

- Extract the types into their own files.
- Hide a bit of the internal state to start to define abstraction boundaries.
  Note that this is pretty minimal so far, just accessors etc, so to keep
  this PR mechanical in nature.