Deprecate `HttpContextKeys.HTTP_TARGET_ADDRESS_BEHIND_PROXY`
Motivation:

Prior to `ConnectionObserver.ProxyConnectObserver` added in #2711, the
sequence of `ConnectionObserver` events was different between regular
connections and connection to the proxy tunnels.
`HttpContextKeys.HTTP_TARGET_ADDRESS_BEHIND_PROXY` (#2169) was useful to
distinguish when the correct callback inside `ConnectionObserver` before
reporting that a new connection is ready to take traffic.
After adding `ProxyConnectObserver`, there is no need in this custom
key.

Modifications:

- Deprecate `HttpContextKeys.HTTP_TARGET_ADDRESS_BEHIND_PROXY` and
describe what is an appropriate replacement for users;

Result:

We will be able to remove
`HttpContextKeys.HTTP_TARGET_ADDRESS_BEHIND_PROXY` in the future
releases.