ReplayStrategies.LazyTimeLimitedReplayAccumulator trims and accumulate
Motivation:

The lazy accumulator doesn't eagerly evict items so it can be prone
to poor behaviors under certain patterns. One of those patterns is
having a large queue limit with a short timeout and infrequent
re-subscribes. In this case a producer can fill the queue to the brim
despite many elements being expired.

Modifications:

Trim stale entries when accumulating.

Result:

This removes a pathological case and leaves the only remaining
pathological case being one where we get a large burst of messages,
none of which have expired, and then we have no more new elements or
subscribers for long durations which can put unnecessary pressure on
the GC.