Fix/Debug Flaky CancellationTest
Motivation
----------
The CancellationTest#cancelOffload() is reportedly flaky and this
PR is a first attempt to understand why / fix the issues.

Modifications
-------------
Investigation found that it looks like on CI the sleeping of the
action races with the cancellation of the service single. Current
working theory is that if the cancellation arrives before the
service is executed, the thread sleep will never be triggered
and so the test completes fine. If the cancellation does not
arrive first, the sleep starts to happen and when the cancellation
arrives afterwards the thread is interrupted as part of the cancellation
which triggers the exception.

Current theory is if this is acceptable behavior, the error
is harmless and can be ignored.