Mark HTTP/2 connection as closing on exception caught (#2686)
Motivation:

Similar to #2675, `H2ParentConnectionContext.AbstractH2ParentConnection` should mark the `Channel` as closing on `exceptionCaught`. If there is no active `Subscriber`, it should enforce channel closure using `ChannelCloseUtils`.
In case users don't have offloading, there is a risk to retry on the same IO thread. We should notify `LoadBalancer` that this connection is closing to avoid retrying on the same connection.

Modifications:

- Invoke `parentContext.notifyOnClosingImpl()` inside `exceptionCaught` before notifying `transportError` and failing subscriber;
- Close the h2 channel if there is no active `Subscriber` to consume the error;
- Remove `H2ParentConnectionContext.hasSubscriber()` that does not seem required anymore;
- Enhance `ConnectionClosedAfterIoExceptionTest` to make sure we fail only the first connect attempt, we capture `RetryableException`s, and capture the callers stack trace if `client.request(...)` fails;

Result:

Retries always select a different connection if existing connection observes an exception.
Fixes #2685.