Fix the flaky LingeringRoundRobinLoadBalancerTest
Motivation:

The test `expiringAHostDoesntRaceWithConnectionAdding` is flaky
but only in CI. This is because locally we never hit the `else`
branch where the connection was added first but it does happen
in CI. That branch of the test is flawed: the host is marked
as EXPIRED and thus unhealthy, so we should always expect a
NoActiveHostException instead of a NoAvailableHostException.

Note that this branch can be tested by wating for the future `f`
to complete before sending the service discovery event.

Modifications:

Change the expected exception to NoActiveHostException.

Result:

Test shouldn't be flaky anymore.