Fix Flaky CancellationTest (#2731)
Motivation
-----------
The CancellationTest#cancelOffload() is flaky and this PR attempts to solve it.

Modifications
--------------
It looks like that on CI the sleeping of the action races with the cancellation of the service Single. If the cancellation arrives before the service is executed, the thread sleep will never be triggered and so the test completes fine. If the cancellation does not arrive first, the sleep starts to happen and when the cancellation arrives afterwards the thread is interrupted as part of the cancellation which triggers the exception.

Since interrupts are expected on cancellations, this exception should be caught and ignored since it is legitimate.