Do not consume orphaned message bodies, just warn.
Motivation
----------
In a previous changeset, the HttpMessageDiscardWatchdogServiceFilter
has been introduced which tries to proactively clean up orphaned/
discarded message body buffers: This works in some, but not all cases
in a reliable fashion.

Modifications
-------------
Instead of trying to clean up the buffers (which might not work all
the time), just WARN into the logs so users can take proactive
action to fix their code.

The log level has been dropped from ERROR to WARN since it is not
a fatal error but still needs to be taken seriously.