Warn if response payloads are not drained properly (#2710)
Motivation
----------
When an exception is raised during the filter pipeline, it is expected that the message bodies
are cleaned up by the filter. If this is not happening, WARN logs should be raised to point
at the issue.

Modifications
-------------
Similar to the service-side filter, this client-side filter tracks the message payloads and warns if 
they are not freed properly. 

Result
------
Proper warning of leaking/non-discarding message payloads.