loadbalancer: Fix asymptotic behavior of host updates
Motivation:

Our current host update process is quadratic both in
time and space due to searching for hosts using list
iteration and performing a COW for each entry of the
SD event set.

Modifications:

Make a Map<Address, SDEvent>. This lets us iterate
through existing events with O(M+N) complexity
instead of O(M*N) complexity. As a bonus, we can
now easily detect if we get multiple SD events for
the same address in a batch.

Result:

We should see slightly worse performance in the small
M*N case but dramatically better performance in the
large case. Since SD events shouldn't be terribly
common and making a hashmap is cheap relative to the
total cost of a SD update, this is the safer
tradeoff.