Fix bug in http/grpc logging observer that prevented request Throwables to be logged
Motivation:

LoggingHttpLifecycleObserver and LoggingGrpcLifecycleObserver included a bug that prevented request exceptions to be logged.

Modifications:

- Update LoggingHttpLifecycleObserver and LoggingGrpcLifecycleObserver to include the Throwable as the last argument in the logger call
- Add tests for the two components

Result:

The Logger object will be called with a Throwable as last argument instead of null in case of request errors. Most logging frameworks implementing slf4j will print the full stacktraces in this situation.