loadbalancer: Add a P2C host selector (#2743)
Motivation

We have a round-robin selector but also want to support
a form of random selection. P2C is a common random
selection algorithm that gives us the ability to also distinguish
hosts based on a load metric, or score, which can help us
distribute load more evenly and avoid hosts that can be
considered outliers by a variety of mechanisms.

Modifications

- Make `Host` an interface and rename the existing version
  to `DefaultHost` to make testing easier and the contract
  explicit. This doesn't mean the contract won't change as we
  continue to iterate.
- Add the `P2CSelector` type and it's tests.

Note: we are not yet integrating this selection algorithm into the
`NewRoundRobinLoadBalancer` which will entail a bunch of
questions around the API for doing so. That will be done as a
followup to minimize the patch size.